debug Sink tinyb
